V1.0.1.10 - first fully functional version
This is the first fully functional version, with original intended features functioning as expected. The only known issue occurs when initially selecting the USB interface for the amplifier. DTR and RTS are toggled when the port selection dialog is selected. This may be an issue if you use either (or both) signals to key the transmitter or amplifier. These signals are toggled on ALL connected USB devices, so be certain that undesired toggling of these signals doesn't make bad things happen! The easiest way to reset the signal is to re-run rigctl with the DTR and RTS toggled off.
Installation (on Ubuntu/Debian, at least) - drop AmpCtl in /usr/local/sbin (or somewhere in your system path), add a shortcut on the desktop or in the menu and give it a spin! 
Can't speak to other distros or versions, this was developed on Ubuntu 18.04 LTS using Lazarus 2.04.