Add files via upload
V1.0.4.1 - fully functional beta release. 
Executable for Linux (Ubuntu 18.04, but should work on others) and Windows (tested on W10). Requires/depends on rigctld running and presenting access to the radio on localhost:4532 for the automatic update from the radio to function. Address and port are changeable if you download the sources and recompile, of course. 
The desktop control and status will function without the automatic update.